feat: Divisão dos componentes do Front pra reaproveitamento e versionamento
