+++ lib/ui/components/action_tile.dart
@@ -1,0 +1,68 @@
+import 'package:flutter/material.dart';
+import '../../theme/colors.dart';
+
+class ActionTile extends StatelessWidget {
+  final String title;
+  final IconData icon;
+  final VoidCallback onTap;
+  final double width;
+
+  const ActionTile({
+    super.key,
+    required this.title,
+    required this.icon,
+    required this.onTap,
+    required this.width,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return SizedBox(
+      width: width,
+      child: ElevatedButton(
+        style: ElevatedButton.styleFrom(
+          elevation: 0,
+          backgroundColor: kCardBg,
+          foregroundColor: const Color(0xFF101828),
+          minimumSize: const Size.fromHeight(64),
+          padding: const EdgeInsets.symmetric(horizontal: 14, vertical: 14),
+          shape: RoundedRectangleBorder(
+            borderRadius: BorderRadius.circular(16),
+            side: const BorderSide(color: kCardBorder, width: 2),
+          ),
+          shadowColor: Colors.black12,
+        ),
+        onPressed: onTap,
+        child: Row(
+          children: [
+            Container(
+              width: 44,
+              height: 44,
+              decoration: BoxDecoration(
+                color: Colors.white,
+                border: Border.all(color: kCardBorder),
+                borderRadius: BorderRadius.circular(12),
+              ),
+              alignment: Alignment.center,
+              child: Icon(icon, size: 24, color: kBrand),
+            ),
+            const SizedBox(width: 12),
+            Expanded(
+              child: Text(
+                title,
+                maxLines: 1,
+                overflow: TextOverflow.ellipsis,
+                style: const TextStyle(
+                  fontWeight: FontWeight.w800,
+                  fontSize: 16,
+                  color: Color(0xFF101828),
+                ),
+              ),
+            ),
+            const Icon(Icons.chevron_right, color: kBrand),
+          ],
+        ),
+      ),
+    );
+  }
+}
