+++ lib/ui/components/loading_placeholder.dart
@@ -1,0 +1,25 @@
+import 'package:flutter/material.dart';
+
+class LoadingPlaceholder extends StatelessWidget {
+  final double height;
+
+  const LoadingPlaceholder({super.key, this.height = 64});
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      height: height,
+      decoration: BoxDecoration(
+        color: const Color(0xFFEFF1F3),
+        borderRadius: BorderRadius.circular(12),
+      ),
+      child: const Center(
+        child: SizedBox(
+          width: 20,
+          height: 20,
+          child: CircularProgressIndicator(strokeWidth: 2),
+        ),
+      ),
+    );
+  }
+}
