+++ lib/ui/components/logo_action.dart
@@ -1,0 +1,33 @@
+import 'package:flutter/material.dart';
+
+class LogoAction extends StatelessWidget {
+  final String imagePath;
+  final double size;
+  final double borderRadius;
+
+  const LogoAction({
+    super.key,
+    required this.imagePath,
+    this.size = 28,
+    this.borderRadius = 6,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.only(right: 4),
+      child: SizedBox(
+        width: size,
+        height: size,
+        child: ClipRRect(
+          borderRadius: BorderRadius.circular(borderRadius),
+          child: Image.asset(
+            imagePath,
+            fit: BoxFit.cover,
+            filterQuality: FilterQuality.medium,
+          ),
+        ),
+      ),
+    );
+  }
+}
