+++ pubspec.yaml
@@ -23,6 +23,8 @@ dependencies:
   font_awesome_flutter:
   fbdb: ^1.5.1
   dio: 5.9.0
+  intl: ^0.19.0
+
 
 
 dev_dependencies:
